fctl/nav: Extract NavLink component
