+++ clients/ctl/admin-ui/src/features/common/nav/NavBar.tsx
@@ -1,46 +1,9 @@
 import { Flex } from "@fidesui/react";
-import NextLink from "next/link";
-import { useRouter } from "next/router";
-import React, { ReactElement } from "react";
+import React from "react";
 
 import { ArrowDownLineIcon } from "~/features/common/Icon";
 
-import { NavButton } from "./NavButton";
-
-interface NavLinkProps {
-  title: string;
-  href: string;
-  disabled?: boolean;
-  rightIcon?: ReactElement;
-  exact?: boolean;
-}
-
-const NavLink = ({ title, href, disabled, rightIcon, exact }: NavLinkProps) => {
-  const router = useRouter();
-  let isActive = false;
-  if (exact) {
-    isActive = router.pathname === href;
-  } else {
-    isActive = router.pathname.startsWith(href);
-  }
-
-  if (disabled) {
-    return (
-      <NavButton
-        disabled={disabled}
-        isActive={isActive}
-        rightIcon={rightIcon}
-        title={title}
-      />
-    );
-  }
-
-  return (
-    <NextLink href={href} passHref>
-      <NavButton isActive={isActive} rightIcon={rightIcon} title={title} />
-    </NextLink>
-  );
-};
+import { NavLink } from "./NavLink";
 
 const NavBar = () => (
   <Flex borderBottom="1px" borderTop="1px" px={9} py={1} borderColor="gray.100">
